backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
pydantic-settings==2.7.0
email-validator==2.2.0
redis==5.2.1
stripe==11.3.0
resend==2.6.0
pytest==8.3.4
httpx==0.28.1
